// src/app/model/contact.ts
/**
 * Created by milad on 12/14/17.
 */
export class Contact {

  id: number;
  phoneNumber: string;
  email: string;
  telegramId: string;
  address: string;

  // constructor(
  // phoneNumber:string,
  // email:string,
  // telegramId:string,
  // address:string){
  //     this.phoneNumber=phoneNumber;
  //     this.email=email;
  //     this.telegramId=telegramId;
  //     this.address=address;
  // }
}